test
Refactor salary calculation and constructor calls

- Changed return type of `calcularSalario` from `float` to `double` in `HorasFeriado.cpp` and `HorasOrdinarias.cpp` for improved precision.
- Updated constructor calls in `SubmenuDeducciones.cpp` and `SubmenuIngresos.cpp` to use `Consola()` instead of `consola()`.
- Modified `lanzar` method in `SubmenuDeducciones.cpp` to instantiate `Builder` directly and commented out certain deduction calls for debugging purposes.

+++ proyecto c++/HorasFeriado.cpp
@@ -4,9 +4,9 @@ HorasFeriado::HorasFeriado(int t, Colaborador* c) : Ingresos(c)
 	tipo = t;
 }
 
-float HorasFeriado::calcularSalario() const
+double HorasFeriado::calcularSalario() const
 {
-	float salarioHora = colab->getSalarioBase() / 8;// 5000 , 4 * 
+	double salarioHora = colab->getSalarioBase() / 8;// 5000 , 4 * 
 	if (tipo == 1) {
 		return salarioHora;/*ordinario*/
 	}
